+++ templates/game/src/server/leaderstats.server.cpp
@@ -1,3 +1,4 @@
+#pragma strict
 #include <cluaupp/roblox.hpp>
 
 void CreateLeaderstats(Player* player) {